Fixed slow person table view load
The icons in the table view were being loaded from the server on the UI thread.  Moved that to a background
thread.  Also changed the image to shows the users avatar rather than their map icon.

Also moved to MagicalRecord 2.3-beta.5.  Fixes a bug so that the ObservationPush and AttachmentPush can watch
the default ManagedObjectContext.