MeViewController loading slow
The MeViewController was also loading slow do to avatar being loaded on main thread.  This now
happens in the background

Also fixed a bug in the people table view cell where avatar image was not nulled out if the user
did not have an avatar.

Map callout for users now also displays their avatar.