Re-work how observation attachment segue works
Fix a problem with clicking an observation attachment from the news feed (iPad split view) to quickly.  When this happens the detail pane in
the split view gets screwed up and you cannot navigate back.  Also seemed bad to segue to a view controller that was already visible on the
screen.  3rd the video/audio from a previous attachment click would not stop playing on a new attachment click.  The split view will now check
to see if the image viewer is already the present view of the detail nav controller.  If so the split view will just update the content instead
of lauching another brand new seque for the same controller.  If the view is not present the split view will launch the segue.

Add navigation header to imageviewer view