UTC/local date/time fix
In a few places the UTC date for an observation was diplayed as UTC and should be local.
Are dates are now displayed as local and use a common NSDateFormatter for display.

Also integrated the changes with NSDate+iso8601 category from the sdk to parse/format
NSDate's

fixes #21

Also fixed a bug where the iPad storyboard was segue directly to the disclaimer view before
authenticating the user.  fixes #20