Move logout unwind to initial view controller
Logout was unwinding to the login view controller.  This breaks when a user
launches with a valid token as the login view controller is skipped.  In that
case the unwind does nothing, so logout button in settings does not work.

Move unwind segue code to initial view controller and removed from
login view controller.

fixes #1