Don't ask for location permission until logged in
When we unwind to the initial login view we stop the location manager.  This will ask for location permissions
if the OS has not already asked.  This was happening when creating an account, which is confusing as the user
is not logged in yet. Just dismiss these view controllers instead of unwind to initial.