Disconnected login (#64)
Disconnected login has been improved.  When a user cannot connect to the server to log in, they are allowed to log in offline if they have already authenticated with the server at least once.  When this happens, they are presented with a dialog that tells them they are logging in disconnected and asks them if they would like to continue.  When they log in, a badge is present on the More tab showing them that they are logged in offline.  They can authenticate with the server at any time from the More tab.